Extract runtime doctor diagnostics services
