+++ src/codex_autorunner/adapters/discord/doctor.py
@@ -7,8 +7,8 @@ from pathlib import Path
 from typing import Any, Optional
 
 from ...core.config import HubConfig
+from ...core.diagnostics.types import DoctorCheck
 from ...core.optional_dependencies import missing_optional_dependencies
-from ...core.runtime import DoctorCheck
 from ...voice.config import VoiceConfig
 from ...voice.provider_catalog import (
     local_voice_provider_spec,
