+++ src/codex_autorunner/server.py
@@ -1,6 +1,7 @@
 from importlib import resources
 
-from .core.runtime import LockError, RuntimeContext, clear_stale_lock, doctor
+from .core.diagnostics.repository import doctor
+from .core.runtime import LockError, RuntimeContext, clear_stale_lock
 from .surfaces.web.app import create_hub_app, create_repo_app
 from .surfaces.web.middleware import BasePathRouterMiddleware
